缺陷[#34025]: 【系统管理-配置管理-视频预置位配置】当前没有采集文件，不选择摄像机，点击导出，显示404页面
    [描述说明]:
    [问题原因]:代码逻辑问题
    [解决方法]:修改代码逻辑
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/TCameraPresetService.java
@@ -175,6 +175,14 @@ public class TCameraPresetService {
             cameraIdList = tCameraPresetDao.selectCameraIdList();
         }
         if(cameraIdList != null && cameraIdList.size()>0){
+
+            List<Long> cameraHavePresetList = tCameraPresetDao.selectCameraHavePreset(cameraIdList);
+            if(cameraHavePresetList == null){
+                return "fail";
+            }
+            if(cameraHavePresetList.size() == 0 ){
+                return "fail";
+            }
             //删除zipPath下的所有文件
             try {
                 String cmd= "rm -rf "+zipPath+"/*";
